pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = [{name = "Your Name", email = "you@example.com"}]
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "google-genai>=1.51.0",
    "gunicorn>=23.0.0",
    "openai>=2.8.1",
    "psycopg2-binary>=2.9.11",
    "werkzeug>=3.1.3",
]

[tool.setuptools]
packages = ["app"]
